+++ CHANGELOG.md
@@ -15,7 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Update timestamp in temp folder to hexadecimal.
 - Update test folder to temporarty the folder `RevitTest` to reduce the `PathTooLongException`. (Fix #48)
 - Update to use `ricaun.Revit.UI.Tasks` library to run async tests.
-- Update to use `ricaun.Revit.UI.Busy`library to check if revit is busy.
+- Update to use `ricaun.Revit.UI.Busy` library to check if revit is busy.
 ### Console
 - Throw exception when `RevitTestUtils` is empty.
 ### TestAdapter